requirements.txt
pandas==2.3.3
numpy==2.3.5
matplotlib==3.10.8
langchain==1.1.3
langchain-core==1.1.3
langchain-openai==1.1.1
openai==2.9.0
python-dotenv==1.0.0
